Split StackFrame into a private part and a part shareable with CoreRT
Make StackFrame a partial class with the majority of code shareable
between CoreCLR and CoreRT, just leaving a small CoreCLR-specific bit
in the original file. This is a preparatory step towards fleshing out
.NET Standard 2.0 StackFrame APIs in AOT runtimes (CoreRT / .NET Native).
